requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
arxiv
langchain-text-splitters
lxml
openai
tiktoken
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
beautifulsoup4
streamlit
pypdf2
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
pydantic==1.10.13